+++ jsonschema-maven-plugin/src/main/java/com/github/victools/jsonschema/plugin/maven/SchemaGeneratorMojo.java
@@ -30,6 +30,7 @@ import com.github.victools.jsonschema.module.javax.validation.JavaxValidationMod
 import com.github.victools.jsonschema.module.javax.validation.JavaxValidationOption;
 import com.github.victools.jsonschema.module.swagger15.SwaggerModule;
 import com.github.victools.jsonschema.module.swagger15.SwaggerOption;
+import com.github.victools.jsonschema.module.swagger2.Swagger2Module;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
@@ -361,9 +362,13 @@ public class SchemaGeneratorMojo extends AbstractMojo {
                     this.getLog().debug("- Adding Swagger 1.5 Module");
                     addSwagger15Module(configBuilder, module);
                     break;
+                case "Swagger2":
+                    this.getLog().debug("- Adding Swagger 2.x Module");
+                    addSwagger2Module(configBuilder, module);
+                    break;
                 default:
                     throw new MojoExecutionException("Error: Module does not have a name in "
-                            + "['Jackson', 'JavaxValidation', 'Swagger15'] or does not have a custom classname.");
+                            + "['Jackson', 'JavaxValidation', 'Swagger15', 'Swagger2'] or does not have a custom classname.");
                 }
             }
         }
@@ -417,7 +422,7 @@ public class SchemaGeneratorMojo extends AbstractMojo {
     }
 
     /**
-     * Add the Swagger module to the generator config.
+     * Add the Swagger (1.5) module to the generator config.
      *
      * @param configBuilder The builder on which the config is added
      * @param module        The modules section form the pom
@@ -437,6 +442,17 @@ public class SchemaGeneratorMojo extends AbstractMojo {
             }
             configBuilder.with(new SwaggerModule(swaggerOptions));
         }
+    }
+
+    /**
+     * Add the Swagger (2.x) module to the generator config.
+     *
+     * @param configBuilder The builder on which the config is added
+     * @param module        The modules section form the pom
+     * @throws MojoExecutionException in case of problems
+     */
+    private void addSwagger2Module(SchemaGeneratorConfigBuilder configBuilder, GeneratorModule module) throws MojoExecutionException {
+        configBuilder.with(new Swagger2Module());
     }
 
     /**